requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
bselib==0.0.5
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
discord==1.0.1
discord-pretty-help==1.3.0
discord-py-slash-command==1.1.2
discord.py==1.7.2
dpymenus==2.1.3
emoji==1.2.0
html5lib==1.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
pycodestyle==2.7.0
pylint==2.8.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
